Drive both players from one input timeline, and roll back when wrong
Milestone 6. A session now connects, agrees on the level, exchanges inputs and
recomputes the frames a wrong prediction spoiled.

Every player reads its input from the timeline, player 1 included. Letting the
local player read its pad directly would make its input land a frame earlier
here than on the other machine, and two simulations that disagree by one frame
of charge disagree by about one block of height. Capture sets a flag so the
patch steps aside for the one read that must reach the real pad; without it the
capture would read back what it stored last frame and nobody would move.

Re-simulation reuses the game's own seam. EntityManager.Update walks entities
calling UpdateComponents, so replaying a frame is calling it again -- no
parallel update path to keep in step with the real one. Resimulation marks those
passes: simulation ignores the flag, presentation reads it. Without that split a
rollback over twenty frames plays every landing sound twenty times and writes
the save twenty times.

A peer's pause goes through the game's own PauseManager rather than suppressing
the update from outside. Suppressing it was the first attempt and it is a trap:
it also stops the pause menu and input polling, so a player would be locked in
the session by their partner's pause with no way to leave. GameLoop already
gates all game logic on IsPaused, so reporting the peer's pause through it
borrows the handling the game already has. SetPause is patched rather than the
frame loop polled, because a pause noticed a frame late lets one side advance a
frame the other did not, and a shared simulation does not heal that.

The connection is one menu control reading Open or Leave by state. No
host-or-join question: whoever opens the lobby invites, Steam's overlay is how
the other side accepts, and the distinction never becomes a screen anyone reads.

Known gap, stated in NetplaySession rather than discovered later: shared level
state is not snapshotted. The players and their scoped gimmick state roll back;
a switch mod's own singletons -- which sand is solid, how far a countdown has
run -- do not. On a gimmick map a correction can therefore leave the world and
the players disagreeing. Ordinary maps are unaffected and are where this should
be tried first.

Steam itself is still untested here; that needs two machines.

+++ PlayerSnapshot.cs
@@ -40,7 +40,7 @@ namespace LocalMultiplayerMod
         private bool _cameraSeeded;
         private Dictionary<object, object> _scoped;
 
-        public int Frame { get; private set; }
+        public long Frame { get; private set; }
 
         public int RootCount
         {
@@ -71,7 +71,7 @@ namespace LocalMultiplayerMod
         private static readonly Dictionary<string, FieldInfo> StateFieldCache =
             new Dictionary<string, FieldInfo>();
 
-        public static PlayerSnapshot Capture(PlayerContext context, int frame)
+        public static PlayerSnapshot Capture(PlayerContext context, long frame)
         {
             var snapshot = new PlayerSnapshot();
             if (context == null || !context.IsAlive)
